pyproject.toml
[project]
name = "tinyhands"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9.6"
dependencies = [
    "azure-identity==1.17.1",
    "cryptography~=41.0.4",
    "django-templated-email==3.0.1",
    "Django==4.2.10",
    "django-bootstrap-form==3.4",
    "django-cors-headers==4.3.0",
    "django-braces==1.15",
    "django-extensions==3.2.3",
    "django-extra-views==0.14.0",
    "django-imagekit==5",
    "django-picklefield==3.1",
    "django-storages[azure]==1.14.3",
    "django-webtest==1.9.11",
    "django-widget-tweaks==1.5.0",
    "django-filter==23.3",
    "djangorestframework==3.14.0",
    "drf-jwt~=1.19.2",
    "factory-boy==3.3.0",
    "Faker==19.13.0",
    "fuzzywuzzy==0.18.0",
    "gdata==2.0.18",
    "google-api-python-client==1.7.4",
    "gunicorn==21.2.0",
    "httplib2==0.22.0",
    "ipdb==0.13.13",
    "lxml==4.9.3",
    "matplotlib==3.8.3",
    "numpy==1.26.0",
    "google-oauth==1.0.1",
    "oauth2client==4.1.3",
    "parse==1.19.1",
    "Pillow==10.2.0",
    "preppy==4.2.2",
    "psycopg2==2.9.9",
    "Pygments==2.16.1",
    "PyMuPDF==1.23.5",
    "pyjwt~=1.7.1",
    "python-dateutil==2.7.3",
    "Levenshtein==0.25.1",
    "pytz==2024.1",
    "recordlinkage==0.16",
    "requests~=2.31.0",
    "reportlab==4.1.0",
    "seaborn==0.13.2",
    "six==1.16.0",
    "scikit-learn==1.4.1.post1",
    "toml==0.10.2",
    "Unipath==1.1",
    "z3c.rml==4.4.0",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["application"]
exclude = ["tmp", "backups", "common_sql"]
